move EntityFXPollution class to pollution package
